+++ R/prototype.R
@@ -7,10 +7,10 @@
 #' A [vetiver_model()] object optionally stores an input data prototype for
 #' checking at prediction time.
 #'
-#' - The default for `save_ptype`, `TRUE`, finds an input data prototype (a
+#' - The default for `save_prototype`, `TRUE`, finds an input data prototype (a
 #' zero-row slice of the training data) via [vetiver_ptype()].
-#' - `save_ptype = FALSE` opts out of storing any input data prototype.
-#' - You may pass your own data to `save_ptype`, but be sure to check that it
+#' - `save_prototype = FALSE` opts out of storing any input data prototype.
+#' - You may pass your own data to `save_prototype`, but be sure to check that it
 #' has the same structure as your training data, perhaps with
 #' [hardhat::scream()].
 #'
@@ -18,7 +18,7 @@
 #'
 #' @return A `vetiver_ptype` method returns a zero-row dataframe, and
 #' `vetiver_create_ptype()` returns either such a zero-row dataframe, `NULL`,
-#' or the dataframe passed to `save_ptype`.
+#' or the dataframe passed to `save_prototype`.
 #'
 #' @examples
 #'
@@ -29,12 +29,12 @@
 #' ## calls the right method for `model` via:
 #' vetiver_ptype(cars_lm)
 #'
-#' ## can also turn off `ptype`
+#' ## can also turn off prototype
 #' vetiver_create_ptype(cars_lm, FALSE)
 #' @examplesIf rlang::is_installed("ranger")
 #' ## some models require that you pass in training features
 #' cars_rf <- ranger::ranger(mpg ~ ., data = mtcars)
-#' vetiver_ptype(cars_rf, ptype_data = mtcars[,-1])
+#' vetiver_ptype(cars_rf, prototype_data = mtcars[,-1])
 #'
 #' @rdname vetiver_create_ptype
 #' @export
@@ -50,23 +50,23 @@ vetiver_ptype.default <- function(model, ...) {
 
 #' @rdname vetiver_create_ptype
 #' @export
-vetiver_create_ptype <- function(model, save_ptype, ...) {
-    if (isTRUE(save_ptype)) {
+vetiver_create_ptype <- function(model, save_prototype, ...) {
+    if (isTRUE(save_prototype)) {
         ptype <- vetiver_ptype(model, ...)
-    } else if (isFALSE(save_ptype)) {
+    } else if (isFALSE(save_prototype)) {
         ptype <- NULL
-    } else if (rlang::inherits_any(save_ptype, "data.frame")) {
-        return(save_ptype)
+    } else if (rlang::inherits_any(save_prototype, "data.frame")) {
+        return(save_prototype)
     } else {
-        abort("The `save_ptype` argument must be TRUE, FALSE, or a dataframe.")
+        abort("The `save_prototype` argument must be TRUE, FALSE, or a dataframe.")
     }
     ptype
 }
 
 check_ptype_data <- function(dots, call = rlang::caller_env()) {
-    if (!rlang::has_name(dots, "ptype_data")) {
-        abort(c("No `ptype_data` available to create an input data prototype",
-                "Pass at least one row of training features as `ptype_data`",
+    if (!rlang::has_name(dots, "prototype_data")) {
+        abort(c("No `prototype_data` available to create an input data prototype",
+                "Pass at least one row of training features as `prototype_data`",
                 "See the documentation for `vetiver_ptype()`"),
               call = call)
     }