+++ tests/testthat/test-stacks.R
@@ -36,7 +36,7 @@ test_that("can pin a stacks model", {
         pinned,
         list(
             model = bundle::bundle(butcher::butcher(frog_reg)),
-            ptype = vctrs::vec_ptype(tree_test),
+            prototype = vctrs::vec_ptype(tree_test),
             required_pkgs = c("glmnet", "parsnip", "recipes", "stacks", "stats", "workflows")
         )
     )