+++ tests/testthat/test-xgboost.R
@@ -27,7 +27,7 @@ test_that("can pin an xgboost model", {
         pinned,
         list(
             model = bundle::bundle(cars_xgb),
-            ptype = vctrs::vec_slice(tibble::as_tibble(mtcars[,2:11]), 0),
+            prototype = vctrs::vec_slice(tibble::as_tibble(mtcars[,2:11]), 0),
             required_pkgs = c("xgboost")
         )
     )